We can now log to non-global contexts
This is an architectural change that breaks the ABI (but not the API).
Previously we would only log to a single global context, and we could save a
record-in-progress with asciilog_stash(). This was completely unworkable in a
multi-threaded system where multiple records could be constructed at the same
time.

So we no have the previous API functions that refer to the global context and
also ..._ctx flavors that refer to a particular context. The idiom is

    asciilog_context_t ctx;
    asciilog_context_clear_ctx(&ctx);
    asciilog_set_field_value_ctx__x(5);
    asciilog_set_field_value_ctx__y(6);
    ....
    asciilog_emit_record_ctx(&ctx);

If threading, each thread would have its own context, and
asciilog_emit_record_ctx() knows not to mix incoming records.

This change also obviates the need for the stash() calls, and they have been
removed.