Got rid of the va_args-style printf() when setting field values
I now have a biggest-possible union field to capture any type of field value.
This is passed in to snprintf(). The C API doesn't know the type of the data
anyway, and since I don't touch the bits, it gets interpreted correctly. One
exception to this is that floating-point values are handled differently in the
usual calling convention on x86-64 Linux, so I special-case these. Hopefully
this reduces the CPU overhead