Removed anchor columns
These didn't work the way I wanted. Threading made them break badly, and I need
to do this differently, probably in a fill-filter. Issue that came up:

  main thread:
    while()
    {
      set anchor field

      tell worker thread to do stuff

      set other fields
      emit()
    }

  worker thread:
    while()
    {
      wait until told to do stuff

      create asciilog context
      set fields
      emit()
    }

I wanted the emit() in the worker thread to see the anchor field in the main
thread. However I would often see the 'set anchor field' for the NEXT frame
happen before the worker thread can emit(), so the anchor field it ends up with
would be the one from the next frame. A post-fill filter would only require that
the main thread emit() for the next frame hasn't happened yet, which maybe is
still too brittle. Perhaps the application should be required to set all the
fields it cares about.