Breaking change: by default we select ROWS, not columns
Previously to select rows we would

  asciilog-filter --matches 'x>5'

and to select columns

  asciilog-filter x y z

This patch flips this around to be more like awk, sysdig, tcpdump, etc. Now to
select rows you do

  asciilog-filter 'x>5'

and to select columns you do

  asciilog-filter -p x,y,z

-p is equivalent to --pick and --print.

This patch doesn't update the documentation because I'm not 100% sure I want to
keep this. Yet.