rel()/diff() computed BEFORE matches
This is somewhat convoluted. I want the meaning of rel() and diff() to be
preserved regardless of any early-exit expressions. I.e. this code is
broken:

- if(!matches) { continue }
- print rel() # update internal state

because the internal state will not be updated if(!matches). I thus do
this instead:

- _rel = rel()
- if(!matches) { continue }
- print _rel

This works. But to make it work, I need to pre-compute all the outermost
rel() and diff() expressions. Outermost because rel(rel(x)) should still
work properly