vnl-join --autoprefix handles numerical filenames better
A common special case is that the input files are of the form aaaNNNbbb where
NNN are numbers. If the numbers are 0-padded, the set of NNN could be "01",
"02", "03". In this case the "0" is a common prefix, so it would not be included
in the file labels, which is NOT what you want here: the labels should be "01",
"02", ... not "1", "2". Here I handle this case by removing all trailing digits
from the common prefix