Handling optional-argument cmdline options in a generic way
My coreutils wrappers that take optional arguments (specified as "option:type")
must be given as --option=arg and NOT '--option arg' (because that's how the
coreutils work). Getopt::Long doesn't allow this, so I have to do it myself.
This patch consolidates this logic so that future utilies can easily access it