vnl-uniq works on OSX
It already worked on FreeBSD, but apparently the OSX people are much more firm
that the FreeBSD people in their belief that their tools should suck as much as
possible. The "uniq" tool on OSX doesn't know about the long versions of a few
options (mostly: --unique, --count, --repeated, --skip-fields). My tool now
favor the short versions of the options whenever possible, so things work on GNU
boxes and FreeBSD boxes and OSX boxes.

OSX also refused to acknowledge that somebody could do 'uniq -c -u', and barfs.
Admittedly, this isn't a useful thing to do, but GNU and FreeBSD just let the
user do the not-useful thing. I had to adjust the tests a bit to deal with this