vnl-filter substitutes the columns with a single s// operation
Previously I would process one column at a time, and the intermediate
substitutions could thus then be erroneously found and replaced by subsequent
substitutions. Doing this in one step is probably faster too. The test the check
for the failure that this bug fixes was committed in the previous revision