Fixed vnl-filter --perl --stream --eval
The --stream option means "flush after each print". But the code was calling
flush() explicitly, so --eval would break the --stream: it would be the user's
job to add their own flush(). This patch adds "$|=1" to the preamble in this
case, so the language would add their own implicit flush().

The equivalent was already being done in awk