Python vnlog.slurp() can use a custom dtype and custom columns
This is useful if we want disparate data types, and is more or less required if
some data is non-numerical. This patch takes dtype AND usecols, to pass to
np.loadtxt(). That's temporary. I'm about to get rid of that completely, to be
able to get all the column IDs from the dtype.

Since this patch is temporary, it doesn't have any documentation or tests. Those
are coming with the full implementation, later