Python vnlog.slurp() can infer column indices from the dtype
This fully supports structured dtypes, non-numerical data and all.

slurp() can now take a dtype even to specify a simple type, not just a
structured type.

If taking a STRUCTURED dtype, slurp returns the array only, no keys or
dict_key_index. Because those aren't needed in that case: numpy takes care of
it.

This patch adds the code and the tests (they pass). It does NOT update any
documentation yet.